# server/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost","description":"Generated server url"}],"paths":{"/api/v1/me/profile-image":{"put":{"tags":["me-controller"],"operationId":"confirmProfileImage","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmProfileImage"}}},"required":true},"responses":{"204":{"description":"No Content"}}},"delete":{"tags":["me-controller"],"operationId":"clearProfileImage","responses":{"204":{"description":"No Content"}}}},"/api/v1/me/password":{"put":{"tags":["me-controller"],"operationId":"changePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePassword"}}},"required":true},"responses":{"204":{"description":"No Content"}}}},"/api/v1/tasks":{"get":{"tags":["task-controller"],"operationId":"list","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TaskView"}}}}}}},"post":{"tags":["task-controller"],"operationId":"add","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTask"}}},"required":true},"responses":{"201":{"description":"Created"}}}},"/api/v1/tasks/{taskId}/files":{"get":{"tags":["task-file-controller"],"operationId":"list_1","parameters":[{"name":"taskId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TaskFileView"}}}}}}},"post":{"tags":["task-file-controller"],"operationId":"attach","parameters":[{"name":"taskId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AttachTaskFile"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TaskFileView"}}}}}}},"/api/v1/push/subscription":{"get":{"tags":["push-controller"],"operationId":"state","parameters":[{"name":"endpoint","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PushSubscriptionStateView"}}}}}},"post":{"tags":["push-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterPushSubscription"}}},"required":true},"responses":{"204":{"description":"No Content"}}},"delete":{"tags":["push-controller"],"operationId":"unregister","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnregisterPushSubscription"}}},"required":true},"responses":{"204":{"description":"No Content"}}}},"/api/v1/me/api-token":{"post":{"tags":["me-controller"],"operationId":"issueApiToken","responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IssuedApiToken"}}}}}},"delete":{"tags":["me-controller"],"operationId":"deleteApiToken","responses":{"204":{"description":"No Content"}}}},"/api/v1/auth/signup":{"post":{"tags":["auth-controller"],"operationId":"signup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Signup"}}},"required":true},"responses":{"202":{"description":"코드를 보냈다"}}}},"/api/v1/auth/signup/resend":{"post":{"tags":["auth-controller"],"operationId":"resendSignupCode","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResendCode"}}},"required":true},"responses":{"202":{"description":"Accepted"}}}},"/api/v1/auth/signup/confirm":{"post":{"tags":["auth-controller"],"operationId":"confirmSignup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmSignup"}}},"required":true},"responses":{"201":{"description":"Created"}}}},"/api/v1/auth/password-reset":{"post":{"tags":["auth-controller"],"operationId":"requestPasswordReset","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestPasswordReset"}}},"required":true},"responses":{"202":{"description":"Accepted"}}}},"/api/v1/auth/password-reset/resend":{"post":{"tags":["auth-controller"],"operationId":"resendPasswordResetCode","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResendCode"}}},"required":true},"responses":{"202":{"description":"Accepted"}}}},"/api/v1/auth/password-reset/confirm":{"post":{"tags":["auth-controller"],"operationId":"confirmPasswordReset","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmPasswordReset"}}},"required":true},"responses":{"204":{"description":"No Content"}}}},"/api/v1/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","responses":{"204":{"description":"No Content"}}}},"/api/v1/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Login"}}},"required":true},"responses":{"204":{"description":"No Content"}}}},"/api/v1/attachments/presign":{"post":{"tags":["attachment-controller"],"operationId":"presign","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PresignAttachment"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PresignedUpload"}}}}}}},"/api/v1/admin/push/failures/{failureId}/revoke":{"post":{"tags":["admin-push-controller"],"operationId":"revoke","parameters":[{"name":"failureId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":"No Content"}}}},"/api/v1/admin/push/failures/{failureId}/resolve":{"post":{"tags":["admin-push-controller"],"operationId":"resolve","parameters":[{"name":"failureId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":"No Content"}}}},"/api/v1/tasks/{taskId}":{"get":{"tags":["task-controller"],"operationId":"detail","parameters":[{"name":"taskId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TaskView"}}}}}},"delete":{"tags":["task-controller"],"operationId":"remove","parameters":[{"name":"taskId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":"No Content"}}},"patch":{"tags":["task-controller"],"operationId":"edit","parameters":[{"name":"taskId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EditTask"}}},"required":true},"responses":{"204":{"description":"No Content"}}}},"/api/v1/tasks/{taskId}/my-day":{"patch":{"tags":["task-controller"],"operationId":"changeMyDay","parameters":[{"name":"taskId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeMyDay"}}},"required":true},"responses":{"204":{"description":"No Content"}}}},"/api/v1/tasks/{taskId}/importance":{"patch":{"tags":["task-controller"],"operationId":"changeImportance","parameters":[{"name":"taskId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeImportance"}}},"required":true},"responses":{"204":{"description":"No Content"}}}},"/api/v1/tasks/{taskId}/completion":{"patch":{"tags":["task-controller"],"operationId":"changeCompletion","parameters":[{"name":"taskId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeCompletion"}}},"required":true},"responses":{"204":{"description":"No Content"}}}},"/api/v1/me":{"get":{"tags":["me-controller"],"operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MeView"}}}}}},"patch":{"tags":["me-controller"],"operationId":"changeNickname","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeNickname"}}},"required":true},"responses":{"204":{"description":"No Content"}}}},"/api/v1/push/config":{"get":{"tags":["push-controller"],"operationId":"config","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PushConfigView"}}}}}}},"/api/v1/admin/users":{"get":{"tags":["admin-user-controller"],"operationId":"list_2","parameters":[{"name":"keyword","in":"query","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AdminUserPageView"}}}}}}},"/api/v1/admin/push/failures":{"get":{"tags":["admin-push-controller"],"operationId":"list_3","parameters":[{"name":"includeResolved","in":"query","required":false,"schema":{"type":"boolean","default":false}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PushFailurePageView"}}}}}}},"/api/v1/tasks/{taskId}/files/{taskFileId}":{"delete":{"tags":["task-file-controller"],"operationId":"detach","parameters":[{"name":"taskId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"taskFileId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":"No Content"}}}}},"components":{"schemas":{"ConfirmProfileImage":{"type":"object","properties":{"objectKey":{"type":"string","maxLength":400,"minLength":0}},"required":["objectKey"]},"ChangePassword":{"type":"object","properties":{"currentPassword":{"type":"string","maxLength":72,"minLength":0},"newPassword":{"type":"string","maxLength":72,"minLength":8}},"required":["currentPassword","newPassword"]},"CreateTask":{"type":"object","properties":{"title":{"type":"string","maxLength":200,"minLength":0},"dueDate":{"type":["string","null"],"format":"date"}},"required":["title"]},"AttachTaskFile":{"type":"object","properties":{"objectKey":{"type":"string","maxLength":400,"minLength":0},"fileName":{"type":"string","maxLength":255,"minLength":0},"contentType":{"type":"string","maxLength":100,"minLength":0}},"required":["contentType","fileName","objectKey"]},"TaskFileView":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"fileName":{"type":"string"},"contentType":{"type":"string"},"size":{"type":"integer","format":"int64"},"url":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}},"required":["contentType","createdAt","fileName","id","size","url"]},"RegisterPushSubscription":{"type":"object","properties":{"endpoint":{"type":"string","maxLength":1000,"minLength":0},"p256dh":{"type":"string","maxLength":255,"minLength":0},"auth":{"type":"string","maxLength":255,"minLength":0}},"required":["auth","endpoint","p256dh"]},"IssuedApiToken":{"type":"object","properties":{"token":{"type":"string"},"issuedAt":{"type":"string","format":"date-time"}},"required":["issuedAt","token"]},"Signup":{"type":"object","properties":{"email":{"type":"string","maxLength":320,"minLength":0},"password":{"type":"string","maxLength":72,"minLength":8},"nickname":{"type":["string","null"],"maxLength":30,"minLength":0}},"required":["email","password"]},"ResendCode":{"type":"object","properties":{"email":{"type":"string","maxLength":320,"minLength":0}},"required":["email"]},"ConfirmSignup":{"type":"object","properties":{"email":{"type":"string","maxLength":320,"minLength":0},"code":{"type":"string","maxLength":16,"minLength":0}},"required":["code","email"]},"RequestPasswordReset":{"type":"object","properties":{"email":{"type":"string","maxLength":320,"minLength":0}},"required":["email"]},"ConfirmPasswordReset":{"type":"object","properties":{"email":{"type":"string","maxLength":320,"minLength":0},"code":{"type":"string","maxLength":16,"minLength":0},"newPassword":{"type":"string","maxLength":72,"minLength":8}},"required":["code","email","newPassword"]},"Login":{"type":"object","properties":{"email":{"type":"string","maxLength":320,"minLength":0},"password":{"type":"string","maxLength":72,"minLength":0}},"required":["email","password"]},"PresignAttachment":{"type":"object","properties":{"slot":{"type":"string","enum":["TASK_FILES","USER_PROFILE_IMAGE"]},"fileName":{"type":"string","maxLength":255,"minLength":0},"contentType":{"type":"string","maxLength":100,"minLength":0},"size":{"type":"integer","format":"int64"}},"required":["contentType","fileName","size","slot"]},"PresignedUpload":{"type":"object","properties":{"objectKey":{"type":"string"},"url":{"type":"string"}},"required":["objectKey","url"]},"EditTask":{"type":"object","properties":{"title":{"type":"string","maxLength":200,"minLength":0},"note":{"type":"string","maxLength":2000,"minLength":0},"dueDate":{"type":["string","null"],"format":"date"},"remindAt":{"type":["string","null"],"format":"date-time","example":"2026-08-30T09:00"}},"required":["dueDate","note","remindAt","title"]},"ChangeMyDay":{"type":"object","properties":{"inMyDay":{"type":"boolean"}},"required":["inMyDay"]},"ChangeImportance":{"type":"object","properties":{"important":{"type":"boolean"}},"required":["important"]},"ChangeCompletion":{"type":"object","properties":{"completed":{"type":"boolean"}},"required":["completed"]},"ChangeNickname":{"type":"object","properties":{"nickname":{"type":"string","maxLength":30,"minLength":0}},"required":["nickname"]},"TaskView":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"title":{"type":"string"},"note":{"type":"string"},"dueDate":{"type":["string","null"],"format":"date"},"remindAt":{"type":["string","null"],"format":"date-time","example":"2026-08-30T09:00"},"important":{"type":"boolean"},"myDayOn":{"type":["string","null"],"format":"date"},"completedAt":{"type":["string","null"],"format":"date-time"},"createdAt":{"type":"string","format":"date-time"}},"required":["completedAt","createdAt","dueDate","id","important","myDayOn","note","remindAt","title"]},"PushSubscriptionStateView":{"type":"object","properties":{"registered":{"type":"boolean"}},"required":["registered"]},"PushConfigView":{"type":"object","properties":{"publicKey":{"type":"string"}},"required":["publicKey"]},"MeView":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"email":{"type":"string"},"nickname":{"type":"string"},"role":{"type":"string","enum":["USER","ADMIN"]},"profileImageUrl":{"type":["string","null"]},"apiTokenIssuedAt":{"type":["string","null"],"format":"date-time"},"createdAt":{"type":"string","format":"date-time"}},"required":["apiTokenIssuedAt","createdAt","email","id","nickname","profileImageUrl","role"]},"AdminUserPageView":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/AdminUserView"}},"total":{"type":"integer","format":"int64"},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"}},"required":["items","page","size","total"]},"AdminUserView":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"email":{"type":"string"},"nickname":{"type":"string"},"role":{"type":"string","enum":["USER","ADMIN"]},"createdAt":{"type":"string","format":"date-time"}},"required":["createdAt","email","id","nickname","role"]},"PushFailurePageView":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/PushFailureView"}},"total":{"type":"integer","format":"int64"},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"}},"required":["items","page","size","total"]},"PushFailureView":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"userId":{"type":"string","format":"uuid"},"email":{"type":["string","null"],"description":"지워진 사용자이면 없다"},"nickname":{"type":["string","null"]},"endpoint":{"type":"string"},"taskId":{"type":["string","null"],"format":"uuid"},"reason":{"type":"string","enum":["FAILED","GONE"]},"detail":{"type":["string","null"]},"occurredAt":{"type":"string","format":"date-time"},"resolvedAt":{"type":["string","null"],"format":"date-time"}},"required":["detail","email","endpoint","id","nickname","occurredAt","reason","resolvedAt","taskId","userId"]},"UnregisterPushSubscription":{"type":"object","properties":{"endpoint":{"type":"string","maxLength":1000,"minLength":0}},"required":["endpoint"]}}}}